make alignForward/alignBackward functions reusable
